requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.5
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
appdirs==1.4.4
appnope==0.1.4
arch==7.2.0
astropy==7.0.0
astropy-iers-data==0.2024.12.9.0.36.21
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.0
camelot-py==0.11.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
clarabel==0.9.0
cleo==0.8.1
click==8.1.7
clikit==0.6.2
comm==0.2.2
contourpy==1.3.0
crashtest==0.3.1
cryptography==44.0.0
cvxpy==1.6.0
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.6
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
ecos==2.0.14
emec-api==0.1.10
entrypoints==0.4
environ==1.0
et-xmlfile==1.1.0
executing==2.1.0
Faker==33.1.0
fastjsonschema==2.21.1
favicon==0.7.0
fire==0.7.0
fonttools==4.54.1
fredapi==0.5.2
frozendict==2.4.6
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.67.0rc1
groq==0.13.1
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
hiredis==3.1.0
htbuilder==0.7.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib-metadata==6.11.0
inflection==0.5.1
investpy==1.0.8
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.7
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-groq==0.2.2
langchain-text-splitters==0.3.4
langsmith==0.2.6
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==3.0.2
marshmallow==3.23.3
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.1
more-itertools==10.5.0
MouseInfo==0.1.3
mplcyberpunk==0.7.1
multidict==6.1.0
multitasking==0.0.11
mypy-extensions==1.0.0
mysql-connector-python==9.0.0
narwhals==1.15.1
Nasdaq-Data-Link==1.0.4
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
num2words==0.5.14
numpy==1.26.4
numpy-financial==1.0.0
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
orjson==3.10.12
osqp==0.6.7.post3
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.3
pandas-datareader==0.10.0
pandocfilters==1.5.1
parso==0.8.4
pastel==0.2.1
patsy==1.0.1
pdf2docx==0.5.8
pdfminer.six==20231228
pdfplumber==0.11.5
peewee==3.17.7
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
plotly-express==0.4.1
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.26.0
protobuf==5.29.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyaml==25.1.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pybind11==2.13.6
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
pyerfa==2.0.1.5
pyfiglet==0.8.post1
PyGetWindow==0.0.9
Pygments==2.18.0
pylev==1.4.0
pymdown-extensions==10.12
PyMsgBox==1.0.9
PyMuPDF==1.25.3
pyobjc-core==10.3.2
pyobjc-framework-Cocoa==10.3.2
pyobjc-framework-Quartz==10.3.2
pyparsing==3.2.0
pypdf==5.1.0
pypdfium2==4.30.1
pyperclip==1.9.0
pyportfolioopt==1.5.6
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-bcb==0.3.0
python-coinmarketcap==0.5
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytweening==1.2.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
qdldl==0.1.7.post4
Quandl==3.7.0
RapidFuzz==3.10.1
redis==5.2.1
referencing==0.35.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.9.4
Riskfolio-Lib==6.3.1
rpds-py==0.21.0
rsa==4.9
rubicon-objc==0.4.9
scikit-learn==1.6.0
scikit-optimize==0.10.2
scipy==1.14.1
scs==3.2.7
seaborn==0.13.2
selenium==4.27.1
setuptools==75.3.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.35
st-annotated-text==4.0.1
stack-data==0.6.3
statsmodels==0.14.4
streamlit==1.29.0
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.3.5
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.2.0
streamlit-lottie==0.0.5
streamlit-option-menu==0.3.6
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tabula-py==2.10.0
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.5.0
thefuzz==0.20.0
threadpoolctl==3.5.0
tinycss2==1.4.0
tinydb==4.8.2
toml==0.10.2
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
unicodecsv==0.14.1
Unidecode==1.3.7
uritemplate==4.1.1
urllib3==2.2.3
validators==0.34.0
watchdog==6.0.0
wcwidth==0.2.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
wsproto==1.2.0
xlrd==2.0.1
XlsxWriter==3.2.0
yarl==1.17.2
yfinance==0.2.37
youtube-transcript-api==0.6.3
zipp==3.21.0
